RF: Move anything needing wx from pkgtools to plugin_manager.utils
