+++ psychopy/app/builder/builder.py
@@ -25,6 +25,7 @@ from wx.lib import scrolledpanel
 from wx.lib import platebtn
 from wx.html import HtmlWindow
 
+import psychopy.app.plugin_manager.dialog
 from .validators import WarningManager
 from ..pavlovia_ui import sync
 from ..pavlovia_ui.project import ProjectFrame
@@ -1385,7 +1386,7 @@ class BuilderFrame(BaseAuiFrame, handlers.ThemeMixin):
             return True
 
     def openPluginManager(self, evt=None):
-        dlg = plugin_manager.EnvironmentManagerDlg(self)
+        dlg = psychopy.app.plugin_manager.dialog.EnvironmentManagerDlg(self)
         dlg.ShowModal()
         # Do post-close checks
         dlg.onClose()
@@ -2859,7 +2860,7 @@ class ComponentsPanel(scrolledpanel.ScrolledPanel, handlers.ThemeMixin):
         dlg.ShowModal()
 
     def onPluginBtn(self, evt=None):
-        dlg = plugin_manager.EnvironmentManagerDlg(self)
+        dlg = psychopy.app.plugin_manager.dialog.EnvironmentManagerDlg(self)
         dlg.ShowModal()
         # Do post-close checks
         dlg.onClose()
